@@ -1,12 +1,11 @@
 package com.atguigu.ssyx.activity.mapper;
 
-
 import com.atguigu.ssyx.model.activity.ActivityInfo;
 import com.atguigu.ssyx.model.activity.ActivityRule;
+import com.atguigu.ssyx.model.activity.ActivitySku;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
-import org.springframework.stereotype.Repository;
+import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.List;
 
@@ -16,12 +15,16 @@ import java.util.List;
  * </p>
  *
  * @author atguigu
- * @since 2024-04-25
+ * @since 2023-04-07
  */
-@Repository
 public interface ActivityInfoMapper extends BaseMapper<ActivityInfo> {
 
-    List<Long> selectExistSkuIdList(@Param("skuIdList")List<Long> skuIdList);
+    // 如果之前参加过，活动正在进行中，排除商品
+    List<Long> selectSkuIdListExist(@Param("skuIdList") List<Long> skuIdList);
 
-    List<ActivityRule> selectActivityRuleList(@Param("skuId")Long skuId);
+    //根据skuId进行查询，查询sku对应活动里面规则列表
+    List<ActivityRule> findActivityRule(@Param("skuId") Long skuId);
+
+    //根据所有skuId列表获取参与活动
+    List<ActivitySku> selectCartActivity(@Param("skuIdList") List<Long> skuIdList);
 }
