@@ -1,10 +1,11 @@
 package com.atguigu.ssyx.activity.mapper;
 
-
 import com.atguigu.ssyx.model.activity.CouponInfo;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-import org.springframework.stereotype.Repository;
+import org.apache.ibatis.annotations.Param;
+import org.springframework.web.bind.annotation.PathVariable;
 
+import java.util.List;
 
 /**
  * <p>
@@ -12,9 +13,15 @@ import org.springframework.stereotype.Repository;
  * </p>
  *
  * @author atguigu
- * @since 2024-04-25
+ * @since 2023-04-07
  */
-@Repository
 public interface CouponInfoMapper extends BaseMapper<CouponInfo> {
 
+    //2 根据skuId+userId查询优惠卷信息
+    List<CouponInfo> selectCouponInfoList(@Param("skuId") Long id,
+                                          @Param("categoryId") Long categoryId,
+                                          @Param("userId") Long userId);
+
+    //1 根据userId获取用户全部优惠卷
+    List<CouponInfo> selectCartCouponInfoList(@Param("userId") Long userId);
 }
