@@ -6,6 +6,7 @@ import com.atguigu.ssyx.model.order.CartInfo;
 import com.atguigu.ssyx.model.product.SkuInfo;
 import com.atguigu.ssyx.vo.activity.ActivityRuleVo;
 import com.atguigu.ssyx.vo.order.CartInfoVo;
+import com.atguigu.ssyx.vo.order.OrderConfirmVo;
 import com.baomidou.mybatisplus.core.metadata.IPage;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.IService;
@@ -13,32 +14,40 @@ import com.baomidou.mybatisplus.extension.service.IService;
 import java.util.List;
 import java.util.Map;
 
+/**
+ * <p>
+ * 活动表 服务类
+ * </p>
+ *
+ * @author atguigu
+ * @since 2023-04-07
+ */
 public interface ActivityInfoService extends IService<ActivityInfo> {
 
-    /**
-     * 分页查询
-     * @param pageParam
-     * @return
-     */
+    //列表
     IPage<ActivityInfo> selectPage(Page<ActivityInfo> pageParam);
 
-    /**
-     * 获取活动规则id
-     * @param activityId
-     * @return
-     */
+    //1 根据活动id获取活动规则数据
     Map<String, Object> findActivityRuleList(Long activityId);
 
-    //保存活动规则信息
+    //2 在活动里面添加规则数据
     void saveActivityRule(ActivityRuleVo activityRuleVo);
 
-    //根据关键字获取sku信息列表
+    //3 根据关键字查询匹配sku信息
     List<SkuInfo> findSkuInfoByKeyword(String keyword);
 
-    /**
-     * 根据skuId获取促销规则信息
-     * @param skuId
-     * @return
-     */
-    List<ActivityRule> findActivityRule(Long skuId);
+    //根据skuId列表获取促销信息
+    Map<Long, List<String>> findActivity(List<Long> skuIdList);
+
+    //根据skuID获取营销数据和优惠卷
+    Map<String, Object> findActivityAndCoupon(Long skuId, Long userId);
+
+    //根据skuId获取活动规则数据
+    List<ActivityRule> findActivityRuleBySkuId(Long skuId);
+
+    //获取购物车里面满足条件优惠卷和活动的信息
+    OrderConfirmVo findCartActivityAndCoupon(List<CartInfo> cartInfoList, Long userId);
+
+    //获取购物车对应规则数据
+    List<CartInfoVo> findCartActivityList(List<CartInfo> cartInfoList);
 }
