@@ -4,27 +4,23 @@ import com.atguigu.ssyx.model.search.SkuEs;
 import com.atguigu.ssyx.vo.search.SkuEsQueryVo;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
+
 import java.util.List;
 
-@Service
 public interface SkuService {
 
-    /**
-     * 上架商品列表
-     * @param skuId
-     */
+    //上架
     void upperSku(Long skuId);
 
-    /**
-     * 下架商品列表
-     * @param skuId
-     */
+    //下架
     void lowerSku(Long skuId);
 
-    /**
-     * 获取爆品商品
-     * @param
-     */
+    //获取爆款商品
     List<SkuEs> findHotSkuList();
+
+    //查询分类商品
+    Page<SkuEs> search(Pageable pageable, SkuEsQueryVo skuEsQueryVo);
+
+    //更新商品热度
+    void incrHotScore(Long skuId);
 }
